split up fetching logic
